refactor: eye icon
extracted the component EyeIcon from the hook useEyeIcon to separate the logic from the markup
